# splitter/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pdf-dispatch",
    "version": "1.26.0",
    "description": "Self-hosted PDF splitting service. Watches a folder, splits multi-page\nPDFs on barcode/QR code detection, and routes output documents by trigger code.\n\nConfigurable entirely from the web interface; every setting is also accessible\nvia this REST API, making pdf-dispatch scriptable from Home Assistant, n8n,\nNode-RED, shell scripts, and any HTTP client.\n\n## Authentication\n\nAll endpoints except `/healthz` support two methods (checked in order):\n\n1. **API key** *(recommended for scripts)*: `X-API-Key: <key>` header.  \n   The key is auto-generated at startup and shown in the *API access* section\n   of the web interface. Override with the `API_KEY` environment variable.\n\n2. **HTTP Basic auth**: active only when both `APP_USERNAME` and\n   `APP_PASSWORD` environment variables are set.\n\nIf neither is configured the API is open (trusted network assumption).\nIf `X-API-Key` is present but wrong the request is **always rejected**\n— there is no fall-through to Basic auth.\n",
    "contact": {
      "name": "pdf-dispatch on GitHub",
      "url": "https://github.com/Lheriss/pdf-dispatch"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This pdf-dispatch instance"
    }
  ],
...
  "security": [
    {
      "apiKey": []
    },
    {
      "basicAuth": []
    },
    {}
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Auto-generated 64-character hex key. Visible (and regeneratable) in the *API access* section of the web interface. Override with the `API_KEY` environment variable.\n"
      },
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Active only when both `APP_USERNAME` and `APP_PASSWORD` environment variables are set.\n"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": [
          "ok",
          "error"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "string",
            "description": "Human-readable error description (in the active UI language)."
          }
        }
      },
      "TaskOutput": {
        "type": "object",
        "required": [
          "filename",
          "path",
          "download_url"
        ],
        "properties": {
          "filename": {
            "type": "string",
            "example": "INVOICE_000042.pdf"
          },
          "path": {
            "type": "string",
            "description": "Path relative to DATA_DIR.",
            "example": "output/INVOICE_000042.pdf"
          },
          "download_url": {
            "type": "string",
            "description": "Ready-to-use URL for GET /api/file/{filepath}.",
            "example": "/api/file/output/INVOICE_000042.pdf"
          }
        }
      },
      "Task": {
        "type": "object",
        "required": [
          "id",
          "filename",
          "status",
          "created_at",
          "updated_at",
          "triggers",
          "outputs",
          "docs_count",
          "error"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "34ec8a5a-4c16-4991-854f-8a3d4a08c4f4"
          },
          "filename": {
            "type": "string",
            "example": "batch_scan.pdf"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "success",
              "error"
            ],
            "description": "- `pending`: file saved in INPUT_DIR, watchdog not yet started\n- `processing`: past stabilisation and PDF validity check\n- `success`: all output files written\n- `error`: invalid PDF, non-PDF file, or unexpected exception\n"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-15T10:30:00"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-15T10:30:05"
          },
          "triggers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Trigger codes detected in the source file (populated on success).",
            "example": [
              "INVOICE"
            ]
          },
          "outputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskOutput"
            },
            "description": "Output files produced (populated on success)."
          },
          "docs_count": {
            "type": "integer",
            "description": "Number of output documents produced.",
            "example": 2
          },
          "error": {
            "type": "string",
            "description": "Error description when status is \"error\"; empty otherwise.",
            "example": ""
          },
          "config_override": {
            "type": "object",
            "description": "Per-file configuration override supplied at upload time (if any). Always present; empty object `{}` when no override was provided. Keys match those accepted by POST /api/upload (split_values, separator_placement, subdirs_by_trigger, delete_source, log_verbose).\n",
            "example": {
              "separator_placement": "after",
              "split_values": [
                {
                  "value": "FACTURE",
                  "page_handling": "delete",
                  "case_sensitive": true
                }
              ]
            }
          }
        }
      },
...
      "UploadedFile": {
        "type": "object",
        "required": [
          "filename",
          "task_id"
        ],
        "properties": {
          "filename": {
            "type": "string",
            "description": "Filename as saved in INPUT_DIR (may differ from the original due to collision-avoidance counter suffixing).\n",
            "example": "invoice_001.pdf"
          },
          "task_id": {
            "type": "string",
            "format": "uuid",
            "description": "Use with GET /api/tasks/{task_id} to track processing.",
            "example": "34ec8a5a-4c16-4991-854f-8a3d4a08c4f4"
          },
          "override": {
            "type": "object",
            "description": "Echo of the config_override applied to this file. Present only when at least one override field was provided in the request; omitted otherwise.\n",
            "example": {
              "separator_placement": "after",
              "split_values": [
                {
                  "value": "FACTURE",
                  "page_handling": "delete",
                  "case_sensitive": true
                }
              ]
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/upload": {
      "post": {
        "summary": "Upload PDFs for processing",
        "description": "Receives one or more PDFs, saves them in INPUT_DIR, and returns a\n`task_id` per file for tracking via GET /api/tasks/{task_id}.\n\n## Barcode-free fallback: `trigger`\n\nUse `trigger` (or `default_trigger`) to name files that contain no\nbarcode — without changing the global configuration.\n\n## Per-file configuration override\n\nThe following fields temporarily replace their global counterparts\n**for the processing of files in this request only**. The persistent\nconfiguration is never modified. All fields are optional; omitted\nfields inherit the global value.\n\n| Field | Type | Description |\n|---|---|---|\n| `split_values` | JSON string | Trigger list (array of objects). Empty `[]` = split on every code. |\n| `separator_placement` | string | `\"before\"` or `\"after\"` |\n| `subdirs_by_trigger` | bool string | `\"true\"` / `\"false\"` |\n| `delete_source` | bool string | `\"true\"` / `\"false\"` |\n| `log_verbose` | bool string | `\"true\"` / `\"false\"` |\n\n`split_values` and `trigger` are independent: `split_values` controls\nwhich barcodes trigger a split; `trigger` is the fallback when no code\nis detected.\n\n```bash\n# Override triggers and placement for this upload only\ncurl -F \"file=@batch.pdf\" \\\n     -F 'split_values=[{\"value\":\"FACTURE\",\"page_handling\":\"delete\",\"case_sensitive\":true}]' \\\n     -F \"separator_placement=after\" \\\n     -F \"subdirs_by_trigger=true\" \\\n     -H \"X-API-Key: <key>\" http://nas:5880/api/upload\n```\n\nThe applied override is echoed in each saved-file entry as `\"override\"`\nand stored in the task under `\"config_override\"`.\n",
        "operationId": "uploadFiles",
        "tags": [
          "documents"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Single PDF file (alias for `files`)."
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "One or more PDF files."
                  },
                  "trigger": {
                    "type": "string",
                    "description": "Default trigger code for barcode-free files (alias: `default_trigger`).\n",
                    "example": "INVOICE"
                  },
                  "split_values": {
                    "type": "string",
                    "description": "JSON array of trigger objects. Replaces the global split_values for this upload only. Each object must have `value` (string) and may have `page_handling` (\"keep\"|\"delete\", default \"keep\") and `case_sensitive` (boolean, default true). Empty array `[]` means every detected code triggers a split.\n",
                    "example": "[{\"value\":\"FACTURE\",\"page_handling\":\"delete\",\"case_sensitive\":true}]"
                  },
                  "separator_placement": {
                    "type": "string",
                    "enum": [
                      "before",
                      "after"
                    ],
                    "description": "Overrides separator placement for this upload only.\n"
                  },
                  "subdirs_by_trigger": {
                    "type": "string",
                    "enum": [
                      "true",
                      "false",
                      "1",
                      "0",
                      "yes",
                      "no"
                    ],
                    "description": "Overrides subfolders-by-trigger option."
                  },
                  "delete_source": {
                    "type": "string",
                    "enum": [
                      "true",
                      "false",
                      "1",
                      "0",
                      "yes",
                      "no"
                    ],
                    "description": "Overrides source-archiving option."
                  },
                  "log_verbose": {
                    "type": "string",
                    "enum": [
                      "true",
                      "false",
                      "1",
                      "0",
                      "yes",
                      "no"
                    ],
                    "description": "Overrides verbose-logging option."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload result. `errors` lists per-file validation failures.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "saved": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UploadedFile"
                      }
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                },
                "example": {
                  "ok": true,
                  "saved": [
                    {
                      "filename": "batch.pdf",
                      "task_id": "34ec8a5a-4c16-4991-854f-8a3d4a08c4f4",
                      "override": {
                        "separator_placement": "after",
                        "split_values": [
                          {
                            "value": "FACTURE",
                            "page_handling": "delete",
                            "case_sensitive": true
                          }
                        ]
                      }
                    }
                  ],
                  "errors": []
                }
              }
            }
          },
          "400": {
            "description": "No file provided.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/file/{filepath}": {
      "get": {
        "summary": "Download output file",
        "description": "Serves a file from OUTPUT_DIR. The `path` field returned by\nGET /api/recent and GET /api/tasks can be used directly as `{filepath}`.\n\nAccess is restricted to OUTPUT_DIR — the configuration file and\nin-progress input PDFs are never accessible.\n",
        "operationId": "downloadFile",
        "tags": [
          "documents"
        ],
        "parameters": [
          {
            "name": "filepath",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path relative to DATA_DIR (e.g. `output/INVOICE_001.pdf`).",
            "example": "output/INVOICE_001.pdf"
          },
          {
            "name": "download",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "0",
                "1"
              ],
              "default": "0"
            },
            "description": "`1` forces Content-Disposition: attachment (download dialog)."
          }
        ],
        "responses": {
          "200": {
            "description": "File content with appropriate Content-Type header.",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "403": {
            "description": "Path resolves outside OUTPUT_DIR.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "File not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks": {
      "get": {
        "summary": "List recent upload tasks",
        "description": "Lists tasks created by POST /api/upload, newest first.\n\nTasks are stored in memory (max 200, FIFO eviction). They are **not**\npersisted across restarts. Files dropped directly into INPUT_DIR or\nreceived via email do not generate tasks.\n",
        "operationId": "listTasks",
        "tags": [
          "tasks"
        ],
        "parameters": [
          {
            "name": "n",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 200
            },
            "description": "Maximum number of tasks to return."
          }
        ],
        "responses": {
          "200": {
            "description": "Task list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "tasks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Task"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}": {
      "get": {
        "summary": "Get task status",
        "description": "Returns the status of a specific upload task.\n\n**Typical polling pattern:**\n```python\nimport requests, time\n\nr = requests.post(\"http://nas:5880/api/upload\",\n                  files={\"file\": open(\"scan.pdf\", \"rb\")},\n                  headers={\"X-API-Key\": key})\ntask_id = r.json()[\"saved\"][0][\"task_id\"]\n\nwhile True:\n    task = requests.get(f\"http://nas:5880/api/tasks/{task_id}\",\n                        headers={\"X-API-Key\": key}).json()[\"task\"]\n    if task[\"status\"] in (\"success\", \"error\"):\n        break\n    time.sleep(1)\n\nfor doc in task[\"outputs\"]:\n    print(doc[\"filename\"], doc[\"download_url\"])\n```\n\nProcessing typically takes 2–10 s depending on PDF size and DPI.\n",
        "operationId": "getTask",
        "tags": [
          "tasks"
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Task UUID returned by POST /api/upload."
          }
        ],
        "responses": {
          "200": {
            "description": "Task details.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "task": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Task not found (unknown ID or evicted from memory).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}